+++ Program.cs
@@ -472,6 +472,7 @@ var smartCaptureOptions = builder.Configuration.GetSection(EINVWORLD.Services.Sm
 builder.Services.AddSingleton(smartCaptureOptions);
 builder.Services.AddScoped<EINVWORLD.Services.SmartCapture.SmartCaptureDocumentService>();
 builder.Services.AddScoped<EINVWORLD.Services.SmartCapture.SmartCaptureCompanyHintService>();
+builder.Services.AddScoped<EINVWORLD.Services.SmartCapture.SmartCaptureAutoSubmitEligibilityService>();
 builder.Services.AddScoped<EINVWORLD.Services.Background.ISyncJobHandler, EINVWORLD.Services.SmartCapture.SmartCaptureExtractionJobHandler>();
 builder.Services.AddScoped<EINVWORLD.Services.Background.ISyncJobHandler, EINVWORLD.Services.SmartCapture.SmartCaptureRetentionJobHandler>();
 builder.Services.AddHostedService<EINVWORLD.Services.SmartCapture.SmartCaptureRetentionScheduler>();